+++ pkg/util/log/log_test.go
@@ -9,7 +9,6 @@ import (
 
 	gokitlog "github.com/go-kit/log"
 	"github.com/go-kit/log/level"
-	dslog "github.com/grafana/dskit/log"
 	"github.com/grafana/dskit/server"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/require"
@@ -36,7 +35,7 @@ func ExampleInitLogger() {
 		LogsPerSecondBurst: 4,
 		Registry:           prometheus.NewPedanticRegistry(),
 	}
-	log.InitLogger(&cfg, false, rateLimitedCfg)
+	cfg.Log = log.InitLogger(cfg.LogFormat, cfg.LogLevel, false, rateLimitedCfg)
 
 	for i := 0; i < 1000; i++ {
 		level.Info(log.Logger).Log("msg", "log.Logger", "test", i+1)
@@ -55,36 +54,11 @@ func ExampleInitLogger() {
 	gokitlog.DefaultTimestampUTC = saveTimestamp
 }
 
-// Check that debug lines are correctly filtered out and that rate limit is satisfied.
-func ExampleNewDefaultLogger() {
-	// Kludge a couple of things so we can do tests repeatably.
-	saveStderr := os.Stderr
-	os.Stderr = os.Stdout
-	saveTimestamp := gokitlog.DefaultTimestampUTC
-	gokitlog.DefaultTimestampUTC = gokitlog.TimestampFormat(
-		func() time.Time { return time.Unix(0, 0).UTC() },
-		time.RFC3339Nano,
-	)
-
-	var lvl dslog.Level
-	_ = lvl.Set("info")
-	logger := log.NewDefaultLogger(lvl, dslog.LogfmtFormat)
-
-	level.Info(logger).Log("msg", "log.Logger", "test", 1)
-	level.Debug(logger).Log("msg", "log.Logger", "test", 2)
-
-	// Output:
-	// ts=1970-01-01T00:00:00Z level=info msg=log.Logger test=1
-
-	os.Stderr = saveStderr
-	gokitlog.DefaultTimestampUTC = saveTimestamp
-}
-
 // Check the overhead of debug logging which gets filtered out.
 func BenchmarkDebugLog(b *testing.B) {
 	cfg := server.Config{}
 	require.NoError(b, cfg.LogLevel.Set("info"))
-	log.InitLogger(&cfg, false, log.RateLimitedLoggerCfg{})
+	log.InitLogger(cfg.LogFormat, cfg.LogLevel, false, log.RateLimitedLoggerCfg{})
 	b.ResetTimer()
 	dl := level.Debug(log.Logger)
 	for i := 0; i < b.N; i++ {